Comentários e Autenticação atualizados v2.0

+++ MinhaGab/src/main/java/backendminhagab/example/MinhaGab/repositories/CommentRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import backendminhagab.example.MinhaGab.models.Comentarios;
 
-public interface CommentRepository extends JpaRepository<Comentarios, Long>{
-    List<Comentarios> findByUserId(Long userId);
-    //adicionar coisas a mais se necessário
+public interface CommentRepository extends JpaRepository<Comentarios, Long> {
+
+    List<Comentarios> findByUserId(Integer userId);  // Adicionando este método
 }
